fix all TS errors
