+++ src/types.ts
@@ -1,0 +1,17 @@
+export interface Item {
+  related: string[];
+  seeAlso: string[];
+  tags: string[];
+  audios: any[];
+  images: string[];
+  title: string;
+  description: string;
+  details: string;
+  examples: Example[];
+  id: string;
+}
+
+interface Example {
+  title: string;
+  description: string;
+}
